Read input and output filename from command line
Use argparse to read optional input and output filenames
from the command line. Default to read from STDIN, allowing
for easy command line piping, e.g. from arecord, rather
than having to alter the command and arguments passed to
a subprocess.Popen call.

Note that because a file read from disk might end before
we get a Ctrl-C, we also have to check for this in the read
loop, raise EOFError when we see it, and catch it alongside
KeyboardInterrupt to trigger flushing the write buffer out
to disk.

Move the `arecord` piped command output behavior from the
InputSink class to the argparse section of the main() func.
This lets us simply pass a file object, whether it reads from
an actual file on disk, stdin, or the output of arecord.

The --arecord option overrides the --input_filename option.

n.b. the arecord stuff is untested, since I'm on a Mac.