Remove arecord and all INPUT global dict use
Since we're now reading input from either a command line specified
file (or a Python file object, if calling from code), or the STDIN,
drop the use of the INPUT global dict for specifying input params.

With no input configs hardcoded, there's no reason to call arecord
via a subprocess, since we don't have configs to pass it anyway, and
it's just as easy to call it from the shell and pipe its output
into SNARP's STDIN.

Note that the previous "default" and "podcaster" input configurations
are memorialized in the README.rst documentation file as piped
UNIX shell commands, as well as in two shell scripts in the new scripts/
directory.