Refactor segmentation logic
Refactor the segmentation logic to make it clear when the various
output actions occur. There are four states:

1. State change + newly silent
2. State change + newly audible
3. No state change + currently silent
4. No state change + currently audible

In states (1) and (2), we've just changed state. Regardless of
whether we're newly silent or newly audible, we must emit the current
chunk: In the one case, as a post roll buffer, and in the other,
following any chunks we've buffered for the pre-roll, as the first
"actual" chunk of audible data.

In state (3), we simply push the current chunk into the pre-roll
ring buffer, which will automatically eject old values to keep the
pre-roll buffer size fixed (we're beginning with a pre-roll buffer
size of one chunk; this will change later as we reduce the chunk
size).

In state (4), we're in the midst of an audible segment, and have an
audible chunk, so we just emit it and move on.