Implement bypass audio option
Add CLI option and corresponding remove_silences functionality
to write out to a bypass audio file – an output file which will
receive 100% of samples read from the input file.

This allows for archiving or monitoring all of a live input while
also producing a shortened version with silent periods removed.