Peak to peak range and IQR for silence detection
Rather than specifying absolute sample limits for both the max
and min sample values in a given chunk, specify the maximum
difference – that is, the maximum allowable max_ - min_.

Using this delta helps us avoid problems with DC offset – if a signal
is not centered on the sample range midpoint, we'll still be able
to use the same silence limits, since we're only looking at how much
of the sample range the signal is covering.

For additional robustness, we also have a delta parameter for the
IRQ: The difference between the first and third quartile values
for the sample values in the given chunk.

This helps avoid false results from sporadic noise which might lead
to high max/min values but is less likely to falsely boost the
Q1/Q3 values – since if it did, it would mean a significant number
of samples in the chunk had large absolute value.

Note that since we can expect most audio signals to be fairly
symmetric across the zero point (ignoring any DC offset), the IQR
is effectively measuring the 50th percentile sample absolute values.

At the moment, both the peak and IQR ranges are specified as sample
values relative to a 16 bit sample width, then normalized relative
to the actual sample width at runtime.

Finally, to make this work, we need real sample values from
frame_to_sample: The padding to four byte values has to go.